Refactor application - refactor clean architecture and move logic from controller to service inside application-layer - refactor config in infrastructure layer to inject dependencies to service from infrastructure to application layer - create a command in the app-layer to create a new student - finalize student controller and use student service - add an example file for application.properties - create a data definition file in sql which creates and configures the db with its tables, columns and relations - create a seed.sql file to seed the database with mock data

+++ src/main/java/ch/nicksohl/winslow/application/port/StudentRepositoryInterface.java
@@ -10,7 +10,7 @@ import ch.nicksohl.winslow.domain.student.Student;
 
 public interface StudentRepositoryInterface {
     List<Student> students();
-    Optional<Student> getStudent(Integer studentId);
+    Optional<Student> getStudent(int studentId);
     Student addStudent(Student student); // add or update -> GET or UPDATE
-    void deleteStudent(Long studentId);
+    void deleteStudent(int studentId);
 }
